requirements.txt
PyYAML
colorama
PyQt6